--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -14,7 +14,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions-rs/toolchain@v1
         with:
-          toolchain: 1.47.0
+          toolchain: 1.51.0
           override: true
       - name: cargo fetch
         uses: actions-rs/cargo@v1
@@ -39,7 +39,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions-rs/toolchain@v1
         with:
-          toolchain: 1.47.0
+          toolchain: 1.51.0
           override: true
       # Build benchmarks to prevent bitrot
       - name: Build benchmarks
@@ -49,7 +49,7 @@ jobs:
           args: --benches --all-features
 
   clippy:
-    name: Clippy (1.47.0)
+    name: Clippy (1.51.0)
     timeout-minutes: 30
     runs-on: ubuntu-latest
 
@@ -57,13 +57,13 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions-rs/toolchain@v1
         with:
-          toolchain: 1.47.0
+          toolchain: 1.51.0
           components: clippy
           override: true
       - name: Run clippy
         uses: actions-rs/clippy-check@v1
         with:
-          name: Clippy (1.47.0)
+          name: Clippy (1.51.0)
           token: ${{ secrets.GITHUB_TOKEN }}
           args: --all-features --all-targets -- -D warnings
 
@@ -119,7 +119,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions-rs/toolchain@v1
         with:
-          toolchain: 1.47.0
+          toolchain: 1.51.0
           override: true
       - run: rustup component add rustfmt
       - uses: actions-rs/cargo@v1