--- a/Cargo.toml
+++ b/Cargo.toml
@@ -24,7 +24,7 @@ byteorder = "1"
 subtle = "2.2.1"
 
 [dev-dependencies]
-bls12_381 = { version = "0.4", features = ["bits"] }
+bls12_381 = "0.4"
 criterion = "0.3"
 hex-literal = "0.3"
 rand = "0.8"