--- a/benches/slow.rs
+++ b/benches/slow.rs
@@ -4,7 +4,7 @@ use bellman::{
 };
 use bls12_381::{Bls12, Scalar};
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
-use ff::{Field, PrimeField};
+use ff::{Field, PrimeFieldBits};
 use futures::Future;
 use group::{Curve, Group};
 use pairing::Engine;