MSRV is now 1.51.0.