@str4d 2021-05-21
Before the upgrade (the second commit in this PR):
```
Gnuplot not found, using plotters backend
Benchmarking multiexp/65536: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 28.5s, or reduce sample count to 10.
multiexp/65536          time:   [288.62 ms 291.94 ms 295.83 ms]                           
Found 12 outliers among 100 measurements (12.00%)
  6 (6.00%) high mild
  6 (6.00%) high severe
```
After the upgrade:
```
Benchmarking multiexp/65536: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 20.1s, or reduce sample count to 20.
multiexp/65536          time:   [199.05 ms 199.63 ms 200.24 ms]                           
                        change: [-32.545% -31.619% -30.805%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```